Extract run status into helper
